requirements.txt
datasets==2.19.1
evidently==0.4.16
fastapi==0.103.2
lightgbm==4.3.0
mlflow==2.14.1
numpy==1.26.4
pandas==2.2.2
prefect==2.14.20
griffe==0.38.0
python-dotenv==1.0.1
pyyaml==6.0.1
scikit-learn==1.4.2
sentence-transformers==2.7.0
torch==2.2.2
uvicorn[standard]==0.29.0
